+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/SentryAutoConfigurationTest.kt
@@ -93,12 +93,13 @@ class SentryAutoConfigurationTest {
             "sentry.attach-threads=true",
             "sentry.attach-stacktrace=true",
             "sentry.server-name=host-001",
+            "sentry.exception-resolver-order=100",
             "sentry.proxy.host=example.proxy.com",
             "sentry.proxy.port=8090",
             "sentry.proxy.user=proxy-user",
             "sentry.proxy.pass=proxy-pass"
         ).run {
-            val options = it.getBean(SentryOptions::class.java)
+            val options = it.getBean(SentryProperties::class.java)
             assertThat(options.readTimeoutMillis).isEqualTo(10)
             assertThat(options.shutdownTimeout).isEqualTo(20)
             assertThat(options.flushTimeoutMillis).isEqualTo(30)
@@ -114,6 +115,7 @@ class SentryAutoConfigurationTest {
             assertThat(options.isAttachThreads).isEqualTo(true)
             assertThat(options.isAttachStacktrace).isEqualTo(true)
             assertThat(options.serverName).isEqualTo("host-001")
+            assertThat(options.exceptionResolverOrder).isEqualTo(100)
             assertThat(options.proxy).isNotNull()
             assertThat(options.proxy!!.host).isEqualTo("example.proxy.com")
             assertThat(options.proxy!!.port).isEqualTo("8090")